fix: refactor dto package
