+++ src/main/java/ru/mirea/megatracker/controllers/CoinController.java
@@ -4,9 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import ru.mirea.megatracker.dto.CoinInfoDTO;
-import ru.mirea.megatracker.dto.CoinPriceHistoryDTO;
-import ru.mirea.megatracker.dto.DetailedCoinInfoDTO;
+import ru.mirea.megatracker.dto.coin.CoinInfoDTO;
+import ru.mirea.megatracker.dto.coin.CoinPriceHistoryDTO;
+import ru.mirea.megatracker.dto.coin.DetailedCoinInfoDTO;
 import ru.mirea.megatracker.services.CoinService;
 import ru.mirea.megatracker.util.CoinErrorResponse;
 import ru.mirea.megatracker.util.UserErrorResponse;
