+++ src/main/java/ru/mirea/megatracker/dto/coin/CoinInfoDTO.java
@@ -1,0 +1,14 @@
+package ru.mirea.megatracker.dto.coin;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@Getter
+@Setter
+public class CoinInfoDTO extends BaseCoinDTO {
+
+    private String iconUrl;
+
+}
